fix(api): correct duration_months column type to INT UNSIGNED
Hibernate schema validation failed on startup: the entity's plain
int durationMonths field expects a 4-byte INTEGER column, but the
migration declared it as SMALLINT UNSIGNED. Match the INT UNSIGNED
convention already used for sort_order and similar columns.

+++ services/api/src/main/resources/db/migration/V3__create_delivery_windows_and_subscription_plans.sql
@@ -32,7 +32,7 @@ CREATE TABLE delivery_subscription_plan (
     name_en VARCHAR(160) NOT NULL,
     name_te VARCHAR(160) NULL,
     billing_period VARCHAR(16) NOT NULL,
-    duration_months SMALLINT UNSIGNED NOT NULL,
+    duration_months INT UNSIGNED NOT NULL,
     sort_order INT UNSIGNED NOT NULL DEFAULT 0,
     active BOOLEAN NOT NULL DEFAULT TRUE,
     version BIGINT UNSIGNED NOT NULL DEFAULT 0,
